requirements.txt
aiomultiprocess==0.9.0
cachetools==3.1.1
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dataclasses-json==0.6.3
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
greenlet==3.0.3
httplib2==0.22.0
idna==3.6
isodate==0.6.1
marshmallow==3.20.1
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dotenv==1.0.0
python-slugify==8.0.1
python-youtube==0.9.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.0
scrapetube==2.5.1
six==1.16.0
SQLAlchemy==2.0.23
text-unidecode==1.3
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uritemplate==3.0.1
urllib3==2.1.0
